Moved logic to UserInjector class
